Redirect deleted translations when republishing an edition
https://trello.com/c/ieIgR5mN/488-deleted-translations-not-unpublished-medium

When publishing an edition, compare the list of currently attached
translations to the list from the previous edition. For each
translation that has been removed, unpublish it with a redirect to the
english version of the edition.

Updates the tests to instantiate a document before passing into
publishing_api_pusher since we'd never expect otherwise (it's
unrealistic data) and you can't test if something has a previous
edition without a document.